refactor(brett): mannequin.ts → mannequin-physics + mannequin-visuals (T000661)
